requirements.txt
Flask==2.3.2
speedtest-cli==2.1.3
